+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.1.15
+
+- Readme: update screenshots in docs
+
 # 0.1.14
 
 - Render environment variables and mappings with a table